// src/payload-types.ts
export interface Media {
  id: string;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "pages".
 */
export interface Page {
  id: string;
  title: string;
  slug: string;
  layout: (
    | {
        title: string;
        highlightedText: string;
        subheading: string;
        tagline?: string | null;
        howIWork?:
          | {
              point: string;
              id?: string | null;
              blockName?: string | null;
              blockType: 'howIWork';
            }[]
          | null;
        resume?: (string | null) | Media;
        id?: string | null;
        blockName?: string | null;
        blockType: 'hero';
      }
    | {
        brands?:
          | {
              brand?: (string | null) | Media;
              id?: string | null;
              blockName?: string | null;
              blockType: 'brand';
            }[]
          | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'brands';
      }
    | {
        aboutVideoLink: string;
        aboutVideoThumb: string | Media;
        id?: string | null;
        blockName?: string | null;
        blockType: 'about';
      }
    | {
        portfolio?:
          | {
              media: string | Media;
              thumbnail?: (string | null) | Media;
              title: string;
              description: {
                root: {
                  type: string;
                  children: {
                    type: string;
                    version: number;
                    [k: string]: unknown;
                  }[];
                  direction: ('ltr' | 'rtl') | null;
                  format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
                  indent: number;
                  version: number;
                };
                [k: string]: unknown;
              };
              mediaType?: string | null;
              link: string;
              client?: string | null;
              year?: number | null;
              id?: string | null;
              blockName?: string | null;
              blockType: 'project';
            }[]
          | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'portfolio';
      }
  )[];
  updatedAt: string;
  createdAt: string;
}